+++ CHANGELOG.md
@@ -43,9 +43,10 @@ The main work (all changes without a GitHub username in brackets in the below li
     -   Enhancement: Refactored Message size handling to dynamically calculate payload size based on transport capabilities
     -   Enhancement: Refactored and cleanup CASE and PASE and corrected handling in some places
     -   Enhancement: Added BTP Idle timeout as defined in Matter specification
+    -   Enhancement: Enhanced default implementation of GeneralDiagnostics cluster with new convenience methods 
+    -   Enhancement: Many more protocol and functionality syncs with matter specification 1.3
     -   Enhancement/Fix: Several fixes and optimizations in Session and Message Exchange handling
     -   Enhancement/Fix: Adjusted MRP behavior with chip and only use/expect MRP ion unreliable channels (UDP). Fixes BLE commissioning
-    -   Enhancement: Many more protocol and functionality syncs with matter specification 1.3
     -   Fix: Adjusted ValidationErrors to be more specific if they should return "InvalidAction" ot "ConstraintError".
     -   Fix: Adjusted some returned errors to be more specific and to the specification (e.g. InvalidAction instead of Failure)
     -   Fix: Fixed StandaloneAck handling to use an outstanding ack number as piggybacked ack number